+++ client/snapshot/restore.go
@@ -37,7 +37,7 @@ func RestoreSnapshotCmd(appCreator servertypes.AppCreator) *cobra.Command {
 			if err != nil {
 				return err
 			}
-			logger := log.NewLogger(cmd.OutOrStdout())
+			logger := log.NewLogger("snapshot", log.WithConsoleWriter(cmd.OutOrStdout()))
 			app := appCreator(logger, db, nil, ctx.Viper)
 
 			sm := app.SnapshotManager()